Refactor StreetView viewer.
Added some extensibility as to allow future pre-loading on streetview
images (to come with controller movement).

Known Issues:

- Certain StreetView images will not works, due to the parsing
method utilized. Biggest offenders appear to be user submitted images,
which are stored on a seperate server with a sligthly different format.
Perhaps support for this can be added in the near future.